pyproject.toml
[tool.poetry]
name = "aggregator"
version = "0.1.0"
description = "just a news aggregator"
authors = ["Steven Joseph <steven@stevenjoseph.in>"]
license = "GPL"

[tool.poetry.dependencies]
python = "^3.9"
ipdb = "0.13.8"
flake8 = "3.9.2"
yapf = "0.31.0"
django-extensions = "3.1.3"
factory-boy = "3.2.0"
django-nose = "1.4.7"
nose-progressive = "1.5.2"
coverage = "5.5"
pytest = "^6.2.4"
pytz = "2021.1"
Django = "3.2.3"
gunicorn = "20.1.0"
newrelic = "6.4.0.157"
django-dotenv = "1.4.2"
django-3-jet = "1.0.8"
celery = "4.4.7"
psycopg2 = "2.8.6"
redis = "3.5.3"
django-model-utils = "4.1.1"
django-unique-upload = "0.2.1"
django-summernote = "0.8.11.6"
django-celery-beat = "2.0.0"
django-activity-stream = "0.10.0"
django-money = "2.0"
djangorestframework = "3.12.4"
djangorestframework-simplejwt = "4.7.1"
Markdown = "3.3.4"
drf-yasg = "1.20.0"
django-filter = "2.4.0"
django-cors-headers = "3.7.0"
django-rest-passwordreset = "1.1.0"
django-rest-swagger = "2.2.0"
easy-thumbnails = "2.7.1"
django-auto-prefetching = "0.1.10"
social-auth-core = "4.1.0"
social-auth-app-django = "4.0.0"
ipython = "7.24.0"
sentry-sdk = "1.1.0"
django-inlinecss = "0.3.0"
django-storages = "1.11.1"
boto3 = "1.17.84"
django-health-check = "3.16.4"
Scrapy = "^2.5.0"
pytest-django = "^4.4.0"
pymongo = "^3.12.0"
dnspython = "^2.1.0"
pylint = "^2.9.6"
black = "^21.7b0"
billiard = "^3.6.4"
pytest-vcr = "^1.0.2"
readability-lxml = "^0.8.1"
html2text = "^2020.1.16"
crochet = "^2.0.0"
pytest-twisted = "^1.13.3"
django-redis = "^5.0.0"
hiredis = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
